gifs/pyproject.toml
[project]
name = "gifs"
version = "0.1.0"
requires-python = ">=3.8"
dependencies = [
    "transformers>=4.36.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "Pillow>=10.0.0",
    "accelerate>=0.25.0",
    "imagehash>=4.3.2",
]

